+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.191.0
 
+- Refine bind function.
 - Drop media support of Gist for keeping word limits of timelines and threads.
 
 ## 0.190.8